+++ src/platform_impl/ios/view.rs
@@ -519,8 +519,8 @@ pub unsafe fn create_window(
             msg_send![window, setScreen:video_mode.monitor().ui_screen()]
         }
         Some(Fullscreen::Borderless(ref monitor)) => {
-            let uiscreen: id = match monitor {
-                Some(ref monitor) => monitor.ui_screen() as id,
+            let uiscreen: id = match &monitor {
+                Some(monitor) => monitor.ui_screen() as id,
                 None => {
                     let uiscreen: id = msg_send![window, screen];
                     uiscreen