+++ src/platform_impl/ios/window.rs
@@ -196,10 +196,9 @@ impl Inner {
                     let () = msg_send![uiscreen, setCurrentMode: video_mode.video_mode.screen_mode];
                     uiscreen
                 }
-                Some(Fullscreen::Borderless(Some(monitor))) => monitor.ui_screen() as id,
-                Some(Fullscreen::Borderless(None)) => {
-                    self.current_monitor_inner().ui_screen() as id
-                }
+                Some(Fullscreen::Borderless(monitor)) => monitor
+                    .unwrap_or_else(|| self.current_monitor_inner())
+                    .ui_screen() as id,
                 None => {
                     warn!("`Window::set_fullscreen(None)` ignored on iOS");
                     return;