+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - **Breaking:** On Web, the canvas element associated to a `Window` is no longer removed from the DOM when the `Window` is dropped.
 - On Web, `WindowEvent::Resized` is now emitted when `Window::set_inner_size` is called.
 - **Breaking:** `Fullscreen` enum now uses `Option<Borderless>` instead of `Borderless` to pick system preferred monitor.
+- **Breaking:** `Fullscreen` enum now uses `Borderless<Option<MonitorHandle>>` instead of `Borderless<MonitorHandle>` to pick system's preferred monitor.
 
 # 0.22.2 (2020-05-16)